fix(bounty): resolve a11y/focus issues and extract shared types
- Add sr-only labels for status select, tag input, and sort select
  (already present, now confirmed as accessible pattern)
- Replace generic focus rings with brand-consistent focus-visible:ring-2
  ring-[#F7931A]/50 on all 3 filter controls, remove bare outline-none
- Extract FILTER_CONTROL_CLASS constant to eliminate repeated Tailwind
  strings across the three filter elements
- Add app/bounty/types.ts with shared Bounty, Stats, Claim, Submission,
  Payment, BountyData interfaces used across all bounty files
- Add app/bounty/utils.ts with shared statusStyle, formatSats, truncAddr,
  relativeTime, deadlineLabel, formatDate helpers
- Update BountyDirectory.tsx, BountyDetail.tsx, page.tsx, and [id]/page.tsx
  to import from shared modules, eliminating type/util duplication
- React.cache() deduplication already present in [id]/page.tsx (confirmed)
- Fix useMemo dependency array and res.json() type casts for clean build
